requirements.txt
# Dépendances principales (OBLIGATOIRES)
# Compatible Python 3.13
fastapi==0.115.0
feedparser==6.0.11
requests==2.32.3
feedgen==1.0.0
uvicorn[standard]==0.32.0
markdown==3.7
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
jinja2==3.1.4

# Dépendances optionnelles (RECOMMANDÉES)
# Pour anti-Cloudflare léger
cloudscraper==1.2.71

# Pour anti-Cloudflare robuste (RECOMMANDÉ)
playwright==1.48.0

tldextract==5.3.1